// src/types/masterData.ts
export interface OfficeMasterUploadRow {
  circle: string
  region: string
  division: string
  divisional_head: string
  divisional_head_mobile: string
  sub_division: string
  sub_divisional_head: string
  sub_divisional_head_mobile: string
  office_name: string
  alternate_office_name: string
  office_id: string
  email_id: string
  office_type_desc: string
  pincode: string
}